Entities standardized

+++ CloudDomain/Entities/Auditing/FullAuditedEntity.cs
@@ -6,7 +6,7 @@ namespace Pos.Domain.Entities
     /// A shortcut of <see cref="FullAuditedEntity{TPrimaryKey}"/> for most used primary key type (<see cref="int"/>).
     /// </summary>
     [Serializable]
-    public abstract class FullAuditedEntity : FullAuditedEntity<int>, IEntity
+    public abstract class FullAuditedEntity : FullAuditedEntity<long>, IEntity
     {
 
     }
