+++ CloudDomain/Entities/Interface/IAggregateRoot.cs
@@ -2,7 +2,7 @@
 
 namespace Pos.Domain.Entities
 {
-    public interface IAggregateRoot : IAggregateRoot<int>, IEntity
+    public interface IAggregateRoot : IAggregateRoot<long>, IEntity
     {
 
     }
